[Messages/Compose] Move UI-related code to separate classes.
